refactor(card): refactor for clean code style
