--- src/components/Card/CardFooter.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components'
-
-const CardFooter = styled.section`
-  display: flex;
-  justify-content: left;
-`
-
-export default CardFooter
