--- src/components/Card/CardImageContainer.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-const CardImageContainer = styled.div`
-  height: 28rem;
-  display: block;
-  overflow: hidden;
-`
-
-export default CardImageContainer
