--- src/components/Card/CardName.js
@@ -1,12 +1,0 @@
-import styled from 'styled-components'
-
-const CardName = styled.p`
-  text-align: left;
-  font-weight: 600;
-  font-size: 1.6rem;
-  line-height: 1.8rem;
-  margin-top: 1.6rem;
-  margin-bottom: 0.4rem;
-`
-
-export default CardName
