--- src/components/Card/CardSubContainer.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-const CardSubContainer = styled.section`
-  display: flex;
-  justify-content: center;
-  z-index: 0;
-`
-
-export default CardSubContainer
