--- src/components/Card/CardSubtitle.js
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-const CardSubtitle = styled.p`
-  text-align: left;
-  font-weight: normal;
-  font-size: 1.6rem;
-  line-height: 2.4rem;
-  margin-bottom: 1.25rem;
-`
-
-export default CardSubtitle
