--- src/components/Card/StyledEditIcon.js
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-import EditIcon from 'components/EditIcon'
-
-const StyledEditIcon = styled(EditIcon)`
-  margin-left: 0.8rem;
-`
-
-export default StyledEditIcon
