--- src/components/Card/StyledImage.js
@@ -1,14 +1,0 @@
-import styled from 'styled-components'
-import ReactImageAppear from 'react-image-appear'
-
-const StyledImage = styled(ReactImageAppear)`
-  cursor: pointer;
-  position: relative;
-  top: 0;
-  left: 0;
-  object-fit: cover;
-  width: 28rem;
-  height: 28rem;
-`
-
-export default StyledImage
